+++ src/main/java/com/cinchapi/runway/Runway.java
@@ -85,6 +85,21 @@ public final class Runway implements AutoCloseable, DatabaseInterface {
      */
     private static long METADATA_RECORD = -1;
 
+    static {
+        // NOTE: Scanning the classpath adds startup costs proportional to the
+        // number of classes defined. We do this once at startup to minimize the
+        // effect of the cost.
+        hierarchies = HashMultimap.create();
+        Logging.disable(Reflections.class);
+        Reflections.log = null; // turn off reflection logging
+        Reflections reflection = new Reflections(new SubTypesScanner());
+        reflection.getSubTypesOf(Record.class).forEach(type -> {
+            hierarchies.put(type, type);
+            reflection.getSubTypesOf(type)
+                    .forEach(subType -> hierarchies.put(type, subType));
+        });
+    }
+
     /**
      * Return a builder that can be used to precisely configure a {@link Runway}
      * instance.
@@ -155,21 +170,6 @@ public final class Runway implements AutoCloseable, DatabaseInterface {
         return Criteria.where().key(Record.SECTION_KEY)
                 .operator(Operator.EQUALS).value(clazz.getName()).and()
                 .group(criteria).build();
-    }
-
-    static {
-        // NOTE: Scanning the classpath adds startup costs proportional to the
-        // number of classes defined. We do this once at startup to minimize the
-        // effect of the cost.
-        hierarchies = HashMultimap.create();
-        Logging.disable(Reflections.class);
-        Reflections.log = null; // turn off reflection logging
-        Reflections reflection = new Reflections(new SubTypesScanner());
-        reflection.getSubTypesOf(Record.class).forEach(type -> {
-            hierarchies.put(type, type);
-            reflection.getSubTypesOf(type)
-                    .forEach(subType -> hierarchies.put(type, subType));
-        });
     }
 
     /**
@@ -537,8 +537,6 @@ public final class Runway implements AutoCloseable, DatabaseInterface {
 
     }
 
-    // TODO: what about loading a specific record using a parent class?
-
     /**
      * Perform the find operation using the {@code concourse} handler.
      * 
@@ -552,6 +550,8 @@ public final class Runway implements AutoCloseable, DatabaseInterface {
         criteria = ensureClassSpecificCriteria(criteria, clazz);
         return concourse.find(criteria);
     }
+
+    // TODO: what about loading a specific record using a parent class?
 
     /**
      * Return the ids of all the {@code Record}s in the {@code clazz}, using the
@@ -642,6 +642,27 @@ public final class Runway implements AutoCloseable, DatabaseInterface {
             throw CheckedExceptions.wrapAsRuntimeException(e);
         }
         return (T) record;
+    }
+
+    /**
+     * Load a record by {@code id} without knowing its class.
+     * 
+     * @param id
+     * @return the loaded record
+     */
+    <T extends Record> T load(long id) {
+        // TODO: consider making this method public...
+        Concourse connection = connections.request();
+        try {
+            Map<String, Set<Object>> data = connection.select(id);
+            String section = (String) Iterables
+                    .getLast(data.get(Record.SECTION_KEY));
+            Class<T> clazz = Reflection.getClassCasted(section);
+            return load(clazz, id, new TLongObjectHashMap<Record>(), data);
+        }
+        finally {
+            connections.release(connection);
+        }
     }
 
     /**